Move benchmark runner out of lodestar-utils src
