+++ packages/beacon-state-transition/test/perf/phase0/epoch/getAttestationDeltas.perf.ts
@@ -1,4 +1,4 @@
-import {BenchmarkRunner} from "@chainsafe/lodestar-utils";
+import {BenchmarkRunner} from "@chainsafe/lodestar-utils/test_utils/benchmark";
 import {allForks, phase0} from "../../../../src";
 import {generatePerfTestCachedBeaconState, initBLS} from "../../util";
