+++ packages/beacon-state-transition/test/perf/phase0/slot/slots.perf.ts
@@ -1,5 +1,5 @@
 import {generatePerfTestCachedBeaconState, initBLS} from "../../util";
-import {BenchmarkRunner} from "@chainsafe/lodestar-utils";
+import {BenchmarkRunner} from "@chainsafe/lodestar-utils/test_utils/benchmark";
 import {allForks} from "../../../../src";
 
 const testCases = [
