+++ packages/utils/src/index.ts
@@ -13,4 +13,3 @@ export * from "./timeout";
 export * from "./types";
 export * from "./verifyMerkleBranch";
 export * from "./json";
-export * from "./perf/utils";
